requirements.txt
requests>=2.31.0
paho-mqtt>=2.0.0
Flask>=3.0.2
